backend/requirements.txt
# --- Core Framework ---
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
pydantic-settings==2.7.0

# --- LlamaIndex (RAG Orchestration) ---
llama-index==0.12.5
llama-index-vector-stores-qdrant==0.4.3
llama-index-llms-openai-like==0.3.3
llama-index-embeddings-huggingface==0.5.0

# --- Vector DB ---
qdrant-client==1.13.2

# --- PDF Parsing ---
pypdf==5.1.0
pdfplumber==0.11.4

# --- Embeddings ---
sentence-transformers==3.3.1
torch==2.5.1

# --- Utilities ---
python-dotenv==1.0.1
openai==1.59.3
httpx==0.28.1
